refactor: remove v1 useAuth, useAuthWs, useSession impl
