+++ apps/frontend/src/pages/commands/sync.vue
@@ -8,8 +8,8 @@ import { useSyncMetadata } from '../../apis/commands/useSyncMetadata'
 import NeedLogin from '../../components/NeedLogin.vue'
 import ChatSelector from '../../components/sync/ChatSelector.vue'
 import SyncStatus from '../../components/sync/SyncStatus.vue'
+import { useSessionStore } from '../../composables/v2/useSessionV2'
 import { useChats } from '../../store/useChats'
-import { useSessionStore } from '../../store/useSession'
 
 const { t } = useI18n()
 const chatStore = useChats()
@@ -17,8 +17,8 @@ const { chats } = storeToRefs(chatStore)
 const { loadChats } = chatStore
 const { executeChatsSync, currentCommand: chatsSyncCommand, syncProgress: chatsSyncProgress } = useSyncChats()
 const { executeMetadataSync, currentCommand: metadataSyncCommand, syncProgress: metadataSyncProgress } = useSyncMetadata()
-const { checkConnection } = useSessionStore()
-const { isConnected } = storeToRefs(useSessionStore())
+const sessionStore = useSessionStore()
+const { isConnected } = storeToRefs(sessionStore)
 
 const selectedChats = ref<number[]>([])
 const priorities = ref<Record<number, number>>({})
@@ -110,8 +110,7 @@ watch(() => currentCommand.value?.status, (status) => {
 // Lifecycle
 onMounted(async () => {
   await loadChats()
-  const connected = await checkConnection(false)
-  if (!connected)
+  if (!isConnected.value)
     showConnectButton.value = true
 })
 </script>
